Feat: agregar policy para eliminar

+++ app/Http/Controllers/PictureController.php
@@ -182,6 +182,8 @@ class PictureController extends Controller
      */
     public function destroy(Picture $picture)
     {
+        Gate::authorize('delete', $picture);
+
         $picture->delete();
 
         return redirect()->route('picture.index')->with([
